editoast: move SignalCache to infra_cache
