+++ editoast/src/infra_cache/object_cache.rs
@@ -2,10 +2,12 @@ mod buffer_stop_cache;
 mod detector_cache;
 mod electrification_cache;
 mod operational_point_cache;
+mod signal_cache;
 
 pub use buffer_stop_cache::BufferStopCache;
 pub use detector_cache::DetectorCache;
 pub use operational_point_cache::OperationalPointCache;
+pub use signal_cache::SignalCache;
 
 cfg_if! {
     if #[cfg(test)] {
